requirements.txt
Django==3.2.1
djangorestframework==3.12.4
Pillow==8.2.0
django-axes==5.15.0
firebase_admin==5.0.0
psycopg2==2.8.6
django-filter==2.4.0
django-tables2==2.4.0
django-bootstrap4==3.0.1
django-crispy-forms==1.11.2
django-seed==0.2.2